Update DESIGN.md s/engines.pm/packageManager/
#140

I just did a find and replace of "engines.pm to packageManager". Also it seems the version has to be 2.0.0 not ^2.0.0
I don't know whether anything else has changed since this document was written.

Maybe "Envisioned workflow" etc. needs to be changed if it is no longer a proposal and is now released.