fix: use `lstatSync` in `generatePosixLink` (#767)
Using `fs.promises.realpath` in `generatePosixLink` returns the fully resolved, absolute path of the symlink target.

However, the symlink variable used in the comparison is a relative path calculated by `path.relative`. This results in a comparison between an absolute path and a relative path, causing the `currentSymlink !== symlink` check to always evaluate to true.